Изменил классы:  - AdsController - закончил методы:    - - updateComments    - - removeAds    - AdsService - закончил методы:    - - getAdsById    - - deleteAdsById    - - updateCommentInAds В моделях FullAds и Comment:    - - переименовал поля обратно на pk (наверно это сокращение от Primary Key)    - - вернул тип полей pk на Integer    - - в классе-модели Comment убрал аннотацию @Entity - надо решить как мы храним комментарии к объявлениям.

+++ src/main/java/pro/sky/finalprojectsky/repository/AdsRepository.java
@@ -7,8 +7,8 @@ import pro.sky.finalprojectsky.model.FullAds;
 import java.util.List;
 
 @Repository
-public interface AdsRepository extends JpaRepository<FullAds, Long> {
+public interface AdsRepository extends JpaRepository<FullAds, Integer> {
 
     // Метод для получения всех объявлений пользователя
-    List<FullAds> findAllByUserId (Long userId);
+    List<FullAds> findAllByUserId (Integer userId);
 }
